Refactor(facebook): split Facebook converters into organic and marketing classes
