+++ tests/Integration/Conversions/FacebookMarketingMetricIntegrationTest.php
@@ -1,20 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Integration\Conversions;
 
-use Classes\Conversions\FacebookGraphConvert;
+use Classes\Conversions\FacebookMarketingMetricConvert;
 use Doctrine\Common\Collections\ArrayCollection;
 use Entities\Analytics\Account;
 use Entities\Analytics\Campaign;
 use Entities\Analytics\Channeled\ChanneledAccount;
 use Entities\Analytics\Channeled\ChanneledCampaign;
-use Entities\Analytics\Page;
-use Entities\Analytics\Post;
 use Enums\Period;
 use Anibalealvarezs\FacebookGraphApi\Enums\MetricSet;
 use Tests\Integration\BaseIntegrationTestCase;
 
-class FacebookGraphConvertTest extends BaseIntegrationTestCase
+class FacebookMarketingMetricIntegrationTest extends BaseIntegrationTestCase
 {
     public function testAdAccountMetricsTransformsDataCorrectly(): void
     {
@@ -53,8 +53,8 @@ class FacebookGraphConvertTest extends BaseIntegrationTestCase
             ]
         ];
 
-        // 2. Act: Pipe Data through the FacebookGraphConvert tool
-        $collection = FacebookGraphConvert::adAccountMetrics(
+        // 2. Act: Pipe Data through the FacebookMarketingMetricConvert tool
+        $collection = FacebookMarketingMetricConvert::adAccountMetrics(
             rows: $rows,
             logger: null,
             accountEntity: $accountEntity,
@@ -66,9 +66,6 @@ class FacebookGraphConvertTest extends BaseIntegrationTestCase
         // 3. Assert
         $this->assertInstanceOf(ArrayCollection::class, $collection);
         
-        // Row 1 has 4 metrics matching the filter (impressions, clicks, spend, actions).
-        // Row 2 has 4 metrics matching the filter (impressions, clicks, spend, cost_per_unique_outbound_click).
-        // Total metrics should be 8
         $this->assertCount(8, $collection);
 
         $metricsMap = [];
@@ -98,91 +95,7 @@ class FacebookGraphConvertTest extends BaseIntegrationTestCase
 
         // Validate Row 2 assertions
         $this->assertArrayHasKey('25-34_female_cost_per_unique_outbound_click', $metricsMap);
-        // The Conversion script has a specific ternary for cost_per_unique_outbound_click checking array indexing:
         $this->assertEquals($rows[1]['cost_per_unique_outbound_click'][0]['value'], $metricsMap['25-34_female_cost_per_unique_outbound_click']->value); 
-    }
-
-    public function testPageMetricsTransformsDataCorrectly(): void
-    {
-        $pagePlatformId = $this->faker->uuid;
-        $pageEntity = new Page();
-        $pageEntity->addUrl($this->faker->url);
-        $pageEntity->addPlatformId($pagePlatformId);
-        $this->entityManager->persist($pageEntity);
-        $this->entityManager->flush();
-
-        $rows = [
-            [
-                'name' => 'page_impressions',
-                'values' => [
-                    ['value' => $this->faker->numberBetween(100, 10000), 'end_time' => $this->faker->iso8601],
-                    ['value' => $this->faker->numberBetween(100, 10000), 'end_time' => $this->faker->iso8601]
-                ]
-            ]
-        ];
-
-        $collection = FacebookGraphConvert::pageMetrics(
-            rows: $rows,
-            pagePlatformId: $pagePlatformId,
-            postPlatformId: '',
-            logger: null,
-            pageEntity: $pageEntity,
-            postEntity: null,
-            period: Period::Daily
-        );
-
-        $this->assertInstanceOf(ArrayCollection::class, $collection);
-        $this->assertCount(2, $collection);
-    }
-
-    public function testIgAccountMetricsTransformsDataCorrectly(): void
-    {
-        $accountEntity = new Account();
-        $accountEntity->addName($this->faker->name . ' Account');
-        $this->entityManager->persist($accountEntity);
-        
-        $igPlatformId = 'ig_' . $this->faker->numerify('#####');
-        $channeledAccount = new ChanneledAccount();
-        $channeledAccount->addPlatformId($igPlatformId);
-        $channeledAccount->addName($this->faker->userName);
-        $channeledAccount->addType(\Enums\Account::INSTAGRAM);
-        $channeledAccount->addChannel(\Enums\Channel::facebook_organic->value);
-        $channeledAccount->addAccount($accountEntity);
-        $this->entityManager->persist($channeledAccount);
-        $this->entityManager->flush();
-
-        $rows = [
-            [
-                'name' => 'follower_count',
-                'total_value' => [
-                    'breakdowns' => [
-                        [
-                            'dimension_keys' => ['country'],
-                            'results' => [
-                                ['dimension_values' => [$this->faker->countryCode], 'value' => $this->faker->numberBetween(1, 1000)],
-                                ['dimension_values' => [$this->faker->countryCode], 'value' => $this->faker->numberBetween(1, 1000)]
-                            ]
-                        ]
-                    ]
-                ]
-            ],
-            [
-                'name' => 'impressions',
-                'total_value' => ['value' => $this->faker->numberBetween(1, 10000)]
-            ]
-        ];
-
-        $collection = FacebookGraphConvert::igAccountMetrics(
-            rows: $rows,
-            date: $this->faker->date(),
-            pageEntity: null,
-            accountEntity: $accountEntity,
-            channeledAccountEntity: $channeledAccount,
-            logger: null,
-            period: Period::Daily
-        );
-
-        $this->assertCount(3, $collection);
     }
 
     public function testCampaignMetricsTransformsDataCorrectly(): void
@@ -204,7 +117,6 @@ class FacebookGraphConvertTest extends BaseIntegrationTestCase
         $channeledAccount->addName($this->faker->company);
         $channeledAccount->addType(\Enums\Account::META_AD_ACCOUNT);
         $channeledAccount->addChannel(\Enums\Channel::facebook_marketing->value);
-        // Campaign requires account technically in DB relations typically, but we only supply it direct for metric config signature
         
         $this->entityManager->persist($channeledAccount);
         $this->entityManager->persist($channeledCampaign);
@@ -220,7 +132,7 @@ class FacebookGraphConvertTest extends BaseIntegrationTestCase
             ]
         ];
 
-        $collection = FacebookGraphConvert::campaignMetrics(
+        $collection = FacebookMarketingMetricConvert::campaignMetrics(
             rows: $rows,
             logger: null,
             channeledAccountEntity: $channeledAccount,
@@ -229,25 +141,23 @@ class FacebookGraphConvertTest extends BaseIntegrationTestCase
             period: Period::Daily
         );
 
-        $this->assertCount(2, $collection); // Impressions and clicks (actions missing from row)
+        $this->assertCount(2, $collection);
     }
 
     public function testRobustness(): void
     {
-        // 1. AdAccount with empty actions (should default to 0)
         $accountEntity = new Account();
         $accountEntity->addName($this->faker->company);
         $this->entityManager->persist($accountEntity);
         $this->entityManager->flush();
 
         $rows = [['date_start' => $this->faker->date(), 'impressions' => '100', 'actions' => []]];
-        $collection = FacebookGraphConvert::adAccountMetrics($rows, null, $accountEntity, metricSet: MetricSet::BASIC);
-        $this->assertCount(2, $collection); // Impressions and actions
+        $collection = FacebookMarketingMetricConvert::adAccountMetrics($rows, null, $accountEntity, metricSet: MetricSet::BASIC);
+        $this->assertCount(2, $collection);
         $this->assertEquals('100', $collection->first()->value);
 
-        // 2. AdAccount with missing spend (should be skipped by metrics filter typically, but let's check)
         $rows = [['date_start' => $this->faker->date(), 'impressions' => '100']];
-        $collection = FacebookGraphConvert::adAccountMetrics($rows, null, $accountEntity, metricSet: MetricSet::BASIC);
+        $collection = FacebookMarketingMetricConvert::adAccountMetrics($rows, null, $accountEntity, metricSet: MetricSet::BASIC);
         $this->assertCount(1, $collection);
     }
 }
