+++ tests/Unit/Classes/Conversions/FacebookMarketingConvertTest.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Tests\Unit\Classes\Conversions;
 
-use Classes\Conversions\FacebookConvert;
+use Classes\Conversions\FacebookMarketingConvert;
 use Doctrine\Common\Collections\ArrayCollection;
 use Enums\Channel;
 use Tests\Unit\BaseUnitTestCase;
 
-class FacebookConvertTest extends BaseUnitTestCase
+class FacebookMarketingConvertTest extends BaseUnitTestCase
 {
     public function testCampaigns(): void
     {
@@ -32,7 +32,7 @@ class FacebookConvertTest extends BaseUnitTestCase
             ]
         ];
 
-        $result = FacebookConvert::campaigns($data, $channeledAccountId);
+        $result = FacebookMarketingConvert::campaigns($data, $channeledAccountId);
 
         $this->assertInstanceOf(ArrayCollection::class, $result);
         $this->assertCount(1, $result);
@@ -66,7 +66,7 @@ class FacebookConvertTest extends BaseUnitTestCase
             ]
         ];
 
-        $result = FacebookConvert::adsets($data, $channeledAccountId);
+        $result = FacebookMarketingConvert::adsets($data, $channeledAccountId);
 
         $this->assertInstanceOf(ArrayCollection::class, $result);
         $this->assertCount(1, $result);
@@ -96,7 +96,7 @@ class FacebookConvertTest extends BaseUnitTestCase
             ]
         ];
 
-        $result = FacebookConvert::ads($data, $channeledAccountId);
+        $result = FacebookMarketingConvert::ads($data, $channeledAccountId);
 
         $this->assertInstanceOf(ArrayCollection::class, $result);
         $this->assertCount(1, $result);
@@ -105,57 +105,5 @@ class FacebookConvertTest extends BaseUnitTestCase
         $this->assertEquals($platformId, $ad->platformId);
         $this->assertEquals($name, $ad->name);
         $this->assertEquals($adsetId, $ad->channeledAdGroupId);
-    }
-
-    public function testPages(): void
-    {
-        $platformId = '111222333';
-        $name = 'Test Page';
-        $accountId = 1;
-
-        $data = [
-            [
-                'id' => $platformId,
-                'name' => $name,
-                'access_token' => 'dummy_token',
-                'category' => 'Test Category',
-            ]
-        ];
-
-        $result = FacebookConvert::pages($data, $accountId);
-
-        $this->assertInstanceOf(ArrayCollection::class, $result);
-        $this->assertCount(1, $result);
-        $page = $result->first();
-
-        $this->assertEquals($platformId, $page->platformId);
-        $this->assertEquals($name, $page->title);
-        $this->assertEquals($accountId, $page->accountId);
-    }
-
-    public function testPosts(): void
-    {
-        $platformId = '777888999';
-        $message = 'Test Post Message';
-        $pageId = 123;
-        $accountId = 1;
-
-        $data = [
-            [
-                'id' => $platformId,
-                'message' => $message,
-                'created_time' => '2024-01-01T00:00:00+0000',
-                'permalink_url' => 'https://facebook.com/post/1',
-            ]
-        ];
-
-        $result = FacebookConvert::posts($data, $pageId, $accountId);
-
-        $this->assertInstanceOf(ArrayCollection::class, $result);
-        $this->assertCount(1, $result);
-        $post = $result->first();
-
-        $this->assertEquals($platformId, $post->platformId);
-        $this->assertEquals($pageId, $post->pageId);
     }
 }
