+++ tests/Unit/Classes/Conversions/FacebookMarketingMetricConvertTest.php
@@ -1,9 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Unit\Classes\Conversions;
 
-use Classes\Conversions\FacebookGraphConvert;
-use Carbon\Carbon;
+use Classes\Conversions\FacebookMarketingMetricConvert;
 use Doctrine\Common\Collections\ArrayCollection;
 use Entities\Analytics\Account;
 use Entities\Analytics\Campaign;
@@ -11,17 +12,13 @@ use Entities\Analytics\Channeled\ChanneledAccount;
 use Entities\Analytics\Channeled\ChanneledAd;
 use Entities\Analytics\Channeled\ChanneledAdGroup;
 use Entities\Analytics\Channeled\ChanneledCampaign;
-use Entities\Analytics\Page;
-use Entities\Analytics\Post;
-use Enums\Channel;
 use Enums\Period;
 use PHPUnit\Framework\MockObject\MockObject;
 use Tests\Unit\BaseUnitTestCase;
+use Carbon\Carbon;
 
-class FacebookGraphConvertTest extends BaseUnitTestCase
+class FacebookMarketingMetricConvertTest extends BaseUnitTestCase
 {
-    private MockObject|Page $page;
-    private MockObject|Post $post;
     private MockObject|Account $account;
     private MockObject|ChanneledAccount $channeledAccount;
     private MockObject|Campaign $campaign;
@@ -32,11 +29,6 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
     protected function setUp(): void
     {
         parent::setUp();
-        $this->page = $this->createMock(Page::class);
-        $this->page->method('getUrl')->willReturn($this->faker->url);
-
-        $this->post = $this->createMock(Post::class);
-        $this->post->method('getPostId')->willReturn('post' . $this->faker->randomNumber());
 
         $this->account = $this->createMock(Account::class);
         $this->account->method('getName')->willReturn($this->faker->word);
@@ -57,43 +49,6 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
         $this->channeledAd->method('getPlatformId')->willReturn('cad' . $this->faker->randomNumber());
     }
 
-    public function testPageMetrics(): void
-    {
-        $value = $this->faker->numberBetween(1, 1000);
-        $platformId = 'page' . $this->faker->randomNumber();
-        $date = $this->faker->iso8601;
-
-        $rows = [
-            [
-                'name' => 'page_impressions',
-                'values' => [
-                    [
-                        'value' => $value,
-                        'end_time' => $date
-                    ]
-                ]
-            ]
-        ];
-
-        $result = FacebookGraphConvert::pageMetrics(
-            $rows,
-            $platformId,
-            '',
-            null,
-            $this->page,
-            null,
-            Period::Daily
-        );
-
-        $this->assertInstanceOf(ArrayCollection::class, $result);
-        $this->assertCount(1, $result);
-        $metric = $result->first();
-        $this->assertEquals('page_impressions', $metric->name);
-        $this->assertEquals($value, $metric->value);
-        $this->assertEquals($platformId, $metric->platformId);
-        $this->assertEquals(Channel::facebook_organic->value, $metric->channel);
-    }
-
     public function testAdAccountMetrics(): void
     {
         $impressions = $this->faker->numberBetween(1000, 5000);
@@ -112,7 +67,7 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
             ]
         ];
 
-        $result = FacebookGraphConvert::adAccountMetrics(
+        $result = FacebookMarketingMetricConvert::adAccountMetrics(
             $rows,
             null,
             $this->account,
@@ -125,70 +80,6 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
         $this->assertContains($metric->name, ['impressions', 'clicks']);
         $this->assertEquals($platformId, $metric->platformId);
         $this->assertCount(2, $metric->dimensions);
-    }
-
-    public function testIgAccountMetrics(): void
-    {
-        $value = $this->faker->numberBetween(1000, 5000);
-        $date = $this->faker->date();
-
-        $rows = [
-            [
-                'name' => 'impressions',
-                'total_value' => [
-                    'value' => $value
-                ]
-            ]
-        ];
-
-        $result = FacebookGraphConvert::igAccountMetrics(
-            $rows,
-            $date,
-            $this->page,
-            $this->account,
-            $this->channeledAccount,
-            null,
-            Period::Daily
-        );
-
-        $this->assertInstanceOf(ArrayCollection::class, $result);
-        $this->assertCount(1, $result);
-        $metric = $result->first();
-        $this->assertEquals('impressions', $metric->name);
-        $this->assertEquals($value, $metric->value);
-        $this->assertEquals($this->channeledAccount->getPlatformId(), $metric->platformId);
-    }
-
-    public function testIgMediaMetrics(): void
-    {
-        $likes = $this->faker->numberBetween(1, 100);
-        $rows = [
-            [
-                'name' => 'likes',
-                'values' => [
-                    [
-                        'value' => $likes
-                    ]
-                ]
-            ]
-        ];
-
-        $result = FacebookGraphConvert::igMediaMetrics(
-            $rows,
-            $this->page,
-            $this->post,
-            $this->account,
-            $this->channeledAccount,
-            null
-        );
-
-        $this->assertInstanceOf(ArrayCollection::class, $result);
-        $this->assertCount(1, $result);
-        $metric = $result->first();
-        $this->assertEquals('likes', $metric->name);
-        $this->assertEquals($likes, $metric->value);
-        $this->assertEquals($this->post->getPostId(), $metric->platformId);
-        $this->assertEquals(Period::Lifetime->value, $metric->period);
     }
 
     public function testCampaignMetrics(): void
@@ -208,7 +99,7 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
             ]
         ];
 
-        $result = FacebookGraphConvert::campaignMetrics(
+        $result = FacebookMarketingMetricConvert::campaignMetrics(
             $rows,
             null,
             $this->channeledAccount,
@@ -238,7 +129,7 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
             ]
         ];
 
-        $result = FacebookGraphConvert::adsetMetrics(
+        $result = FacebookMarketingMetricConvert::adsetMetrics(
             $rows,
             null,
             $this->channeledAccount,
@@ -269,7 +160,7 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
             ]
         ];
 
-        $result = FacebookGraphConvert::adMetrics(
+        $result = FacebookMarketingMetricConvert::adMetrics(
             $rows,
             null,
             $this->channeledAccount,
@@ -284,29 +175,12 @@ class FacebookGraphConvertTest extends BaseUnitTestCase
         $this->assertCount(3, $result);
     }
 
-    public function testRobustness(): void
+    public function testAdAccountRobustness(): void
     {
-        // 1. Page metrics with missing 'values'
-        $rows = [['name' => 'page_impressions']];
-        $result = FacebookGraphConvert::pageMetrics($rows, 'p123');
-        $this->assertCount(0, $result);
-
-        // 2. Ad account metrics with missing 'date_start'
+        // Ad account metrics with missing 'date_start'
         $rows = [['impressions' => 100]];
-        $result = FacebookGraphConvert::adAccountMetrics($rows, null, $this->account, 'ca123');
+        $result = FacebookMarketingMetricConvert::adAccountMetrics($rows, null, $this->account, 'ca123');
         $this->assertCount(1, $result);
         $this->assertEquals(Carbon::now()->toDateString(), $result->first()->metricDate);
-
-        // 3. IG account metrics with missing 'total_value'
-        $rows = [['name' => 'impressions']];
-        $result = FacebookGraphConvert::igAccountMetrics($rows, '2023-01-01', $this->page, $this->account, $this->channeledAccount);
-        $this->assertCount(1, $result);
-        $this->assertEquals(0, $result->first()->value);
-
-        // 4. IG media metrics with empty 'values'
-        $rows = [['name' => 'likes', 'values' => []]];
-        $result = FacebookGraphConvert::igMediaMetrics($rows, $this->page, $this->post, $this->account, $this->channeledAccount);
-        $this->assertCount(1, $result);
-        $this->assertEquals(0, $result->first()->value);
     }
 }
